Update JDK and Apache Maven
In order to use JDK 9 or later for compiling bspsrc, but still compile
for Java 8, the Apache Maven Compiler Plugin has been configured to use
the '--release' command-line option instead of '-source' and '-target'
with javac. This ensures that the compiled program will not accidentally
use APIs that are only available in the Java version corresponding to
the JDK in use.

The Apache Maven Compiler Plugin and the Apache Maven Shade Plugin have
also been updated to their current versions.